Chore: move dimensions code to virtualization
