+++ public/app/features/logs/components/panel/processing.ts
@@ -3,13 +3,10 @@ import Prism from 'prismjs';
 import { dateTimeFormat, LogLevel, LogRowModel, LogsSortOrder } from '@grafana/data';
 
 import { escapeUnescapedString, sortLogRows } from '../../utils';
-import { LOG_LINE_BODY_FIELD_NAME } from '../LogDetailsBody';
 import { FieldDef, getAllFields } from '../logParser';
 
-import { getDisplayedFieldValue } from './LogLine';
 import { GetFieldLinksFn } from './LogList';
 import { generateLogGrammar } from './grammar';
-import { measureTextWidth } from './virtualization';
 
 export interface LogListModel extends LogRowModel {
   body: string;
@@ -18,11 +15,6 @@ export interface LogListModel extends LogRowModel {
   displayLevel: string;
   fields: FieldDef[];
   timestamp: string;
-}
-
-export interface LogFieldDimension {
-  field: string;
-  width: number;
 }
 
 export interface PreProcessOptions {
@@ -86,47 +78,3 @@ function logLevelToDisplayLevel(level = '') {
       return level;
   }
 }
-
-export const calculateFieldDimensions = (logs: LogListModel[], displayedFields: string[] = []) => {
-  if (!logs.length) {
-    return [];
-  }
-  let timestampWidth = 0;
-  let levelWidth = 0;
-  const fieldWidths: Record<string, number> = {};
-  for (let i = 0; i < logs.length; i++) {
-    let width = measureTextWidth(logs[i].timestamp);
-    if (width > timestampWidth) {
-      timestampWidth = Math.round(width);
-    }
-    width = measureTextWidth(logs[i].displayLevel);
-    if (width > levelWidth) {
-      levelWidth = Math.round(width);
-    }
-    for (const field of displayedFields) {
-      width = measureTextWidth(getDisplayedFieldValue(field, logs[i]));
-      fieldWidths[field] = !fieldWidths[field] || width > fieldWidths[field] ? Math.round(width) : fieldWidths[field];
-    }
-  }
-  const dimensions: LogFieldDimension[] = [
-    {
-      field: 'timestamp',
-      width: timestampWidth,
-    },
-    {
-      field: 'level',
-      width: levelWidth,
-    },
-  ];
-  for (const field in fieldWidths) {
-    // Skip the log line when it's a displayed field
-    if (field === LOG_LINE_BODY_FIELD_NAME) {
-      continue;
-    }
-    dimensions.push({
-      field,
-      width: fieldWidths[field],
-    });
-  }
-  return dimensions;
-};
