+++ public/app/features/logs/components/panel/virtualization.ts
@@ -1,5 +1,7 @@
 import { BusEventWithPayload, GrafanaTheme2 } from '@grafana/data';
 
+import { LOG_LINE_BODY_FIELD_NAME } from '../LogDetailsBody';
+
 import { getDisplayedFieldValue } from './LogLine';
 import { LogListModel } from './processing';
 
@@ -191,6 +193,55 @@ export function getLogLineSize(
   const { height } = measureTextHeight(textToMeasure, getLogContainerWidth(container), optionsWidth);
   return height;
 }
+
+export interface LogFieldDimension {
+  field: string;
+  width: number;
+}
+
+export const calculateFieldDimensions = (logs: LogListModel[], displayedFields: string[] = []) => {
+  if (!logs.length) {
+    return [];
+  }
+  let timestampWidth = 0;
+  let levelWidth = 0;
+  const fieldWidths: Record<string, number> = {};
+  for (let i = 0; i < logs.length; i++) {
+    let width = measureTextWidth(logs[i].timestamp);
+    if (width > timestampWidth) {
+      timestampWidth = Math.round(width);
+    }
+    width = measureTextWidth(logs[i].displayLevel);
+    if (width > levelWidth) {
+      levelWidth = Math.round(width);
+    }
+    for (const field of displayedFields) {
+      width = measureTextWidth(getDisplayedFieldValue(field, logs[i]));
+      fieldWidths[field] = !fieldWidths[field] || width > fieldWidths[field] ? Math.round(width) : fieldWidths[field];
+    }
+  }
+  const dimensions: LogFieldDimension[] = [
+    {
+      field: 'timestamp',
+      width: timestampWidth,
+    },
+    {
+      field: 'level',
+      width: levelWidth,
+    },
+  ];
+  for (const field in fieldWidths) {
+    // Skip the log line when it's a displayed field
+    if (field === LOG_LINE_BODY_FIELD_NAME) {
+      continue;
+    }
+    dimensions.push({
+      field,
+      width: fieldWidths[field],
+    });
+  }
+  return dimensions;
+};
 
 export function hasUnderOrOverflow(element: HTMLDivElement, calculatedHeight?: number): number | null {
   const height = calculatedHeight ?? element.clientHeight;
